requirements.txt
pydantic
langchain
langgraph
langchain-core
langchain-community
langchain-openai
fastapi
uvicorn
PyGithub